Correct solution.
This was a lot harder than I originally expected. The issue is that the
crashes happen also before a cart has had the opportunity to move.

It requires comparing the new position of a cart with all current
positions. This lead to inner loops, one of which needs to mutate the
cart. This was pretty hard in Rust, so I decided to use the RefCell
escape hatch and track crashed carts as a property of the cart.

I'm not very satified with this approach, since the crashed property is
only used within the `tick` method.