Puzzle23 part 2
I have to admit that I cheated for this one. I went to the reddit thread
to get inpiration and found this Z3 solution which was super
interesting.

It's interesting because it solves the problem in a very simple way, but
also as a learning experience since it introduces a crate that has a c
library dependency. I learned more of the rust tooling for those.

Note that using z3 isn't particularly fast, in fact it takes about 5m30s
in my input, but the code is extremely
simple, so the payoff is pretty big.