Puzzle 24 part 2.
There's some edge case where we end up with a group that deals less
damage than the number of unit hit points.

The instructions say that groups only take on whole unit damage and
doesn't say that we should accumulate damage between fights, so I guess
this means we have to detect the case. But the instructions also don't
mention what to do in stalemates...